CHECKPOINT: Yay, this checkpoint deployed spaboilerplate2017.zinglax.com
as of 03/22/2017.  A button was clicked Post Project JSON which in turn called an ajax action.  That action created a SpaBoiler python object capable of producing a new flask site for the Project JSON passed through. After copying the new project from the projects folder on the server, changing a fabfile path, then running fab deploy, a new site with multiple pages can be created.